Removed 1 unnecessary stubbing in FinishReleaseTaskTest.java
